refactor: merge modules reduce files
