+++ nicegold_v5/AGENTS.md
@@ -363,3 +363,6 @@
 - เพิ่มฟังก์ชัน validate_indicator_inputs และเรียกใช้ใน main.py เพื่อตรวจสอบข้อมูลก่อนสร้างสัญญาณ (Patch v11.9.9)
 ### 2025-09-19
 - เพิ่มฟังก์ชัน sanitize_price_columns แปลงคอลัมน์ราคาเป็น float และ log จำนวน NaN พร้อมใช้งานใน main.py (Patch v11.9.10)
+### 2025-09-20
+- รีแฟคเตอร์รวมโมดูล patch และ risk เข้ากับ backtester และลดไฟล์หลักเหลือ 10 ไฟล์
+- รวมชุดทดสอบทั้งหมดเหลือ 3 ไฟล์
