+++ nicegold_v5/changelog.md
@@ -339,3 +339,6 @@
 - เพิ่ม validate_indicator_inputs สำหรับตรวจสอบคอลัมน์และจำนวนแถว และเรียกใช้ใน welcome() (Patch v11.9.9)
 ## 2025-09-19
 - เพิ่ม sanitize_price_columns แปลงคอลัมน์ราคาเป็น float และ log จำนวน NaN พร้อมเรียกใช้ใน main.py (Patch v11.9.10)
+## 2025-09-20
+- รีแฟคเตอร์รวมโค้ด patch และ risk เข้ากับ backtester ลดจำนวนไฟล์หลักเหลือ 10
+- รวมไฟล์ทดสอบทั้งหมดเหลือ 3 ไฟล์
