+++ nicegold_v5/qa.py
@@ -4,6 +4,7 @@ from datetime import datetime
 import pandas as pd
 import numpy as np
 
+# --- QA Guard Functions ---
 
 def detect_overfit_bias(trades: pd.DataFrame) -> dict:
     """Return simple statistics to gauge overfitting."""
@@ -101,3 +102,58 @@ def analyze_drawdown(equity_df: pd.DataFrame) -> dict:
     dd = (equity - peak) / peak
     max_dd = dd.min()
     return {'max_drawdown_pct': round(abs(max_dd) * 100, 2)}
+
+
+# --- Export & Auto QA ---
+
+def export_fold_qa(fold_name: str, stats: dict, bias_score: float, drawdown: dict, outdir: str = "logs"):
+    """Export fold QA summary and related metrics as JSON."""
+    os.makedirs(outdir, exist_ok=True)
+    outpath = os.path.join(outdir, f"fold_qa_{fold_name.lower()}.json")
+    payload = {
+        "fold": fold_name,
+        "qa_summary": stats,
+        "overfit_bias_score": bias_score,
+        "drawdown": drawdown,
+    }
+    with open(outpath, "w") as f:
+        json.dump(payload, f, indent=2, default=lambda o: o.item() if hasattr(o, "item") else o)
+    print(f"📁 Exported QA summary → {outpath}")
+
+
+def detect_fold_drift(fold_stats: list[dict]) -> pd.DataFrame:
+    """Detect metric variations across folds."""
+    df = pd.DataFrame(fold_stats)
+    if len(df) < 2:
+        return pd.DataFrame()
+    metrics = ["winrate", "avg_pnl", "sl_count"]
+    drifts = []
+    for m in metrics:
+        std = df[m].std()
+        mean = df[m].mean()
+        pct_std = round((std / abs(mean)) * 100, 2) if mean != 0 else 0.0
+        drifts.append({"metric": m, "std_dev": round(std, 2), "%_variation": pct_std})
+    return pd.DataFrame(drifts)
+
+
+QA_BASE_PATH = "/content/drive/MyDrive/NICEGOLD/logs/qa"
+os.makedirs(QA_BASE_PATH, exist_ok=True)
+
+
+def auto_qa_after_backtest(trades: pd.DataFrame, equity: pd.DataFrame, label: str = "Run"):
+    """Automatically export QA summary after backtest with timestamp-enhanced label."""
+    if trades.empty:
+        print("❌ ไม่มีข้อมูล trades สำหรับ QA")
+        return
+
+    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+    label_full = f"{label}_{timestamp}"
+
+    stats = summarize_fold(trades, label_full)
+    bias = compute_fold_bias(trades)
+    dd = analyze_drawdown(equity)
+    export_fold_qa(label_full, stats, bias, dd, outdir=QA_BASE_PATH)
+
+    save_csv_path = os.path.join(QA_BASE_PATH, f"fold_qa_{label_full.lower()}.csv")
+    pd.DataFrame([stats | {"bias_score": bias} | dd]).to_csv(save_csv_path, index=False)
+    print(f"✅ QA Auto Export → {save_csv_path}")
